Restructure based on suggestions
