+++ src/assets/favicons/about.txt
