+++ src/assets/favicons/android-chrome-512x512.png
