+++ src/assets/favicons/chromewebstore-dark.png
