+++ src/assets/favicons/github-dark.png
