+++ src/assets/favicons/github-light.png
