+++ src/pages/newtab/art.js
