+++ src/pages/newtab/style.css
