+++ src/shared/defaultSettings.js
